Added missing files renamed
